simplify workout display

+++ Deadlift Diaries/Models/Exercise.swift
@@ -16,10 +16,10 @@ final class Exercise  {
     var weight: Int
     var sets: Int
     var reps: Int
-    var rest: Int
+    var rest: Double
     var creationDate: Double
 
-    init(name: String = "", weight: Int = 0, sets: Int = 0, reps: Int = 0, rest: Int = 0) {
+    init(name: String = "", weight: Int = 0, sets: Int = 0, reps: Int = 0, rest: Double = 0.0) {
         self.id = UUID()
         self.name = name
         self.weight = weight
